Fix Gift Exchange
Was erroring out if during the end phase there were more than 1 continuous effect to resolve and the cards banished to send to the hands were no longer banished

+++ official/c82257940.lua
@@ -47,6 +47,7 @@ function s.thfilter(c,fid)
 end
 function s.thcon(e,tp,eg,ep,ev,re,r,rp)
 	local g=e:GetLabelObject()
+	if g:IsDeleted() then return false end
 	if g:Filter(s.thfilter,nil,e:GetLabel()):GetCount()<2 then
 		g:DeleteGroup()
 		return false
